Move no-thru properties into WayPropertySetSource
