+++ src/test/java/org/opentripplanner/openstreetmap/model/OSMWithTagsTest.java
@@ -3,6 +3,9 @@ package org.opentripplanner.openstreetmap.model;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.graph_builder.module.osm.DefaultWayPropertySetSource;
+import org.opentripplanner.graph_builder.module.osm.WayPropertySet;
+import org.opentripplanner.graph_builder.module.osm.WayPropertySetSource;
 
 public class OSMWithTagsTest {
 
@@ -105,34 +108,5 @@ public class OSMWithTagsTest {
         
         o.addTag("access", "no");
         assertTrue(o.isGeneralAccessDenied());
-    }
-    
-    @Test
-    public void isMotorThroughTrafficExplicitlyDisallowed() {
-        OSMWithTags o = new OSMWithTags();
-        assertFalse(o.isMotorVehicleThroughTrafficExplicitlyDisallowed());
-        
-        o.addTag("access", "something");
-        assertFalse(o.isMotorVehicleThroughTrafficExplicitlyDisallowed());
-        
-        o.addTag("access", "destination");
-        assertTrue(o.isMotorVehicleThroughTrafficExplicitlyDisallowed());
-
-        o.addTag("access", "private");
-        assertTrue(o.isMotorVehicleThroughTrafficExplicitlyDisallowed());
-
-        assertTrue(way("motor_vehicle", "destination").isMotorVehicleThroughTrafficExplicitlyDisallowed());
-    }
-
-    @Test
-    public void isBicycleNoThroughTrafficExplicitlyDisallowed() {
-        assertTrue(way("bicycle", "destination").isBicycleNoThroughTrafficExplicitlyDisallowed());
-        assertTrue(way("access", "destination").isBicycleNoThroughTrafficExplicitlyDisallowed());
-    }
-
-    public  OSMWithTags way(String key, String value) {
-        var way = new OSMWithTags();
-        way.addTag(key, value);
-        return way;
     }
 }
